requirements.txt
streamlit
opencv-python-headless
tensorflow==2.20.0
numpy
pandas
scikit-learn
matplotlib
nltk
spotipy
python-dotenv
pillow
